==> 04_indicator-species-analysis.Rmd <==
---
editor_options: 
  chunk_output_type: console
---

# Indicator species analysis  

This analysis aims to identify what species are “indicators” of groups of samples or treatment type and asks if this varies between point count estimates and acoustic surveys. 

If the aim is to determine which species can be used as indicators of certain site group, an approach commonly used in ecology is the Indicator Value (Dufrene and Legendre, 1997). These authors defined an Indicator Value (IndVal) index to measure the association between a species and a site group. The method of Dufrene and Legendre (1997) calculates the IndVal index between the species and each site group and then looks for the group corresponding to the highest association value. Finally, the statistical significance of this relationship is tested using a permutation test.  

Diagnostic (or indicator) species are an important tool in vegetation science, because these species can be used to characterize and indicate specific plant community types. A statistic commonly used to determine the association (also known as fidelity, not to be confounded with the indicator value component) between species and vegetation types is Pearson’s phi coefficient of association (Chytry et al., 2002). This coefficient is a measure of the correlation between two binary vectors. The abundance-based counterpart of the phi coefficient is called the point biserial correlation coefficient (which is defined as "r.g").  

See: https://cran.r-project.org/web/packages/indicspecies/vignettes/indicspeciesTutorial.pdf  

## Load necessary libraries
```{r}
library(tidyverse)
library(dplyr)
library(stringr)
library(vegan)
library(ggplot2)
library(scico)
library(psych)
library(ecodist)
library(RColorBrewer)
library(ggforce)
library(ggalt)
library(patchwork)
library(sjPlot)
library(indicspecies)

# Source any custom/other internal functions necessary for analysis
source("code/01_internal-functions.R")
```

## Load dataframe containing point count and acoustic data
```{r}
datSubset <- read.csv("results/datSubset.csv")
```

## Load species trait data
```{r}
trait <- read.csv("data/species-trait-dat.csv")

# add it to the subset data
datSubset <- left_join(datSubset,trait[,c(1,7,8)], by = "scientific_name")
```

## Estimate detections across visits for point count data and acoustic data  
These detections are calculate at the site level (for a total of five/six visits for acoustic & point count data, respectively). If the across_visit_detections = 5, that means that a species was detected every single time across each of the five visits to that site. This value ranges from 1 to 5 (or 1 to 6 for point count data). These detections are estimated for the indicator species analysis since the matrix required for the same cannot be a presence/absence matrix.      
```{r}
## we will estimate abundance across point counts by site-date (essentially corresponding to visit)
abundance <- datSubset %>%
  filter(data_type == "point_count") %>%
  group_by(date, site_id, restoration_type, scientific_name,
           common_name, eBird_codes, habitat, foraging_habit) %>% summarise(totAbundance = sum(number)) %>%
  ungroup()

# estimate across visit detections for point count data
pc_visit_detections <- abundance %>%
  mutate(forDetections = case_when(totAbundance > 0 ~ 1)) %>%
  group_by(scientific_name, site_id,restoration_type) %>%
  summarise(across_visit_detections = sum(forDetections)) %>%
  mutate(data_type = "point_count") %>%
  ungroup()

# estimate total number of detections across the acoustic data by site-date (essentially corresponds to a visit)
# note: we cannot call this abundance as it refers to the total number of vocalizations across a 16-min period across all sites
detections <- datSubset %>%
  filter(data_type == "acoustic_data") %>%
  group_by(date, site_id, restoration_type, scientific_name,
           common_name, eBird_codes, habitat, foraging_habit) %>% summarise(totDetections = sum(number)) %>%
  ungroup()

# estimate across visit detections for acoustic data
aru_visit_detections <- detections %>%
  mutate(forDetections = case_when(totDetections > 0 ~ 1)) %>%
  group_by(scientific_name, site_id,restoration_type) %>%
  summarise(across_visit_detections = sum(forDetections)) %>%
  mutate(data_type = "acoustic_data") %>%
  ungroup()
```

## Indicator species analysis of the point-count dataset

The detections across visits is converted to a wide format for the sake of analysis.  
```{r}
## preparing the point count data for the indicator species analysis
pc_indicator <-  pc_visit_detections %>%
  group_by(scientific_name, site_id, restoration_type) %>%
  pivot_wider(names_from = scientific_name, 
               values_from = across_visit_detections, 
               values_fill = list(across_visit_detections=0))

## indicator species analysis
indic_pc <- multipatt(pc_indicator[,4:ncol(pc_indicator)], pc_indicator$restoration_type, func = "r.g", control = how(nperm=999))

# analyze/summarize the results
summary(indic_pc)
```

Multilevel pattern analysis

 Association function: r.g
 Significance level (alpha): 0.05

 Total number of species: 83
 Selected number of species: 29 
 Number of species associated to 1 group: 20 
 Number of species associated to 2 groups: 9 

 List of species associated to each combination: 

 Group AR  #sps.  1 
                       stat p.value  
Zosterops palpebrosus 0.408    0.02 *

 Group BM  #sps.  15 
                               stat p.value    
Alcippe poioicephala          0.851   0.001 ***
Culicicapa ceylonensis        0.737   0.001 ***
Hypothymis azurea             0.636   0.001 ***
Ducula badia                  0.563   0.002 ** 
Cyornis pallidipes            0.556   0.001 ***
Phylloscopus magnirostris     0.553   0.001 ***
Pellorneum ruficeps           0.532   0.001 ***
Harpactes fasciatus           0.492   0.001 ***
Leptocoma minima              0.485   0.002 ** 
Chalcophaps indica            0.462   0.011 *  
Irena puella                  0.451   0.010 ** 
Dicrurus paradiseus           0.435   0.009 ** 
Dicrurus aeneus               0.423   0.008 ** 
Larvivora brunnea             0.375   0.033 *  
Chrysocolaptes guttacristatus 0.356   0.049 *  

 Group NR  #sps.  4 
                        stat p.value    
Cinnyris asiaticus     0.651   0.001 ***
Acrocephalus dumetorum 0.553   0.001 ***
Argya subrufa          0.380   0.031 *  
Machlolophus aplonotus 0.329   0.037 *  

 Group AR+BM  #sps.  3 
                      stat p.value    
Iole indica          0.564   0.001 ***
Muscicapa muttui     0.515   0.002 ** 
Terpsiphone paradisi 0.348   0.050 *  

 Group AR+NR  #sps.  6 
                         stat p.value    
Streptopelia chinensis  0.740   0.001 ***
Pycnonotus jocosus      0.730   0.001 ***
Dicaeum concolor        0.452   0.005 ** 
Psittacula cyanocephala 0.420   0.010 ** 
Psittacula columboides  0.408   0.027 *  
Orthotomus sutorius     0.390   0.021 *  

## Indicator species analysis for acoustic data    

The detections across visits is converted to a wide format for the sake of analysis.  
```{r}
## preparing the acoustic data for the indicator species analysis
aru_indicator <-  aru_visit_detections %>%
  group_by(scientific_name, site_id, restoration_type) %>%
  pivot_wider(names_from = scientific_name, 
               values_from = across_visit_detections, 
               values_fill = list(across_visit_detections=0))

## indicator species analysis
indic_aru <- multipatt(aru_indicator[,4:ncol(aru_indicator)], aru_indicator$restoration_type, func = "r.g", control = how(nperm=999))

# analyze/summarize the results
summary(indic_aru)
```

Multilevel pattern analysis

 Association function: r.g
 Significance level (alpha): 0.05

 Total number of species: 113
 Selected number of species: 32 
 Number of species associated to 1 group: 21 
 Number of species associated to 2 groups: 11 

 List of species associated to each combination: 

 Group AR  #sps.  1 
                    stat p.value   
Halcyon smyrnensis 0.416    0.01 **

 Group BM  #sps.  15 
                               stat p.value    
Alcippe poioicephala          0.702   0.001 ***
Phylloscopus magnirostris     0.697   0.001 ***
Cyornis pallidipes            0.686   0.001 ***
Hypothymis azurea             0.587   0.001 ***
Sitta frontalis               0.582   0.001 ***
Culicicapa ceylonensis        0.547   0.001 ***
Ducula badia                  0.513   0.002 ** 
Leptocoma minima              0.508   0.002 ** 
Dryocopus javensis            0.451   0.003 ** 
Accipiter trivirgatus         0.426   0.033 *  
Harpactes fasciatus           0.401   0.017 *  
Chrysocolaptes guttacristatus 0.388   0.026 *  
Irena puella                  0.381   0.039 *  
Muscicapa muttui              0.379   0.023 *  
Dicrurus paradiseus           0.368   0.049 *  

 Group NR  #sps.  5 
                       stat p.value  
Carpodacus erythrinus 0.407   0.028 *
Pycnonotus cafer      0.387   0.030 *
Prinia hodgsonii      0.378   0.050 *
Treron affinis        0.369   0.029 *
Hierococcyx varius    0.352   0.040 *

 Group AR+NR  #sps.  11 
                         stat p.value    
Pycnonotus jocosus      0.851   0.001 ***
Orthotomus sutorius     0.527   0.001 ***
Streptopelia chinensis  0.508   0.001 ***
Pavo cristatus          0.508   0.002 ** 
Acrocephalus dumetorum  0.475   0.005 ** 
Merops leschenaulti     0.465   0.007 ** 
Psittacula cyanocephala 0.444   0.007 ** 
Zosterops palpebrosus   0.431   0.011 *  
Corvus splendens        0.408   0.012 *  
Copsychus saularis      0.405   0.024 *  
Aegithina tiphia        0.362   0.046 *
