==> 05_repeatability_analysis.Rmd <==
---
editor_options: 
  chunk_output_type: console
---


## Repeatability analysis

In this script, we will plot data on abundance across visits by species (in other words, we choose the 1st, 3rd and 5th visit and compare it with the 2nd, 4th and 6th visits). We will extract the R2 values for the above comparison. We will then repeat the same for acoustic data on detection rates. Following this analysis, we will regress point count data against acoustic detection rate and extract R2 values. 

## Install necessary libraries
```{r}
library(tidyverse)
library(dplyr)
library(stringr)
library(vegan)
library(ggplot2)
library(scico)
library(data.table)
library(extrafont)
library(ggstatsplot)
library(ggside)
library(MASS)
library(scales)
library(rr2)
library(ggnewscale)
library(ggpubr)
library(gridtext)
library(broom)
library(rptR)

# Source any custom/other internal functions necessary for analysis
source("code/01_internal-functions.R")
```

## Load dataframe containing point count and acoustic data
```{r}
datSubset <- read.csv("results/datSubset.csv")
```

## Estimate abundance for point count data and detections for acoustic data  

Here, we make a distinction before running correlative analyses that abundance corresponds to the total number of individuals of a species detected across visits to a site and can only be calculated for point count data. In the acoustic dataset, individuals are not seen and a measure of detections (estimated as the total number of times as species was heard across ~540 10-s clips). Here 576 clips correspond to the total amount of acoustic data - 90 min (540 10-s clips) of data = 15-min of data for every visit). Here, we also add a visit_number column to each object to ease subsampling of data.    
```{r}
# point-count data
# estimate total abundance across all species for each site
# include a vist_number column to sub sample data later on
abundance <- datSubset %>%
  filter(data_type == "point_count") %>%
  group_by(site_id, scientific_name,
           common_name, eBird_codes, date) %>% 
  summarise(abundance_pc = sum(number)) %>%
  mutate(visit_number = row_number()) %>%
  ungroup()

# estimate total number of detections across the acoustic data
# note: we cannot call this abundance as it refers to the total number of vocalizations across all sites
# include a vist_number column to sub sample data later on
detections <- datSubset %>%
  filter(data_type == "acoustic_data") %>%
  group_by(site_id, scientific_name,
           common_name, eBird_codes, date) %>% 
  summarise(detections_aru = sum(number)) %>%
  mutate(visit_number = row_number()) %>%
  ungroup()

# note: we have six visits to each site for point count data but only a maximum of five visits per site for acoustic detections data 

# estimating acoustic detection rates for each species for each site and visit
aruRate <- detections %>%
  mutate(aruRate = detections_aru/90)

# Note that the column aruRate can vary between 0 to 1 for each species for each site (this value can vary across sites for each species, referring to how vocally active a species is)
```

## Run analysis at the level of each species (across all sites and treatments combined) for point count data  

Here we run analyses separately for point count data and acoustic data. First, we will filter data by visits (choosing the 1st, 3rd and 5th visit) to a site and calculate total abundance for each species. We will repeat the above by choosing the 2nd, 4th, and 6th visit to a site and calculate total abundance for each species. 
```{r}
# get rid of species with less than 20 total abundance
spp_subset <-  abundance %>%
  group_by(common_name) %>%
  summarise(abundance_pc = sum(abundance_pc)) %>%
  ungroup() %>%
  filter(abundance_pc >=20)

# subset data
# we have 46 species in total
data <- abundance %>%
  filter(common_name %in% spp_subset$common_name)

# creating empty objects to save outputs
plots <- list()
r2 <- data.frame()

for (i in 1:length(unique(data$common_name))){

 # extract species common name
  a <- unique(data$common_name)[i]
  
  # subset data for plotting
  for_plot <- data[data$common_name==a,]
  
  # grouping point count data
  # 1st, 3rd, and 5th visit
  abundance_groupA <- for_plot %>%
    filter(visit_number %in% c(1,3,5)) %>%
    mutate(abundance_groupA = abundance_pc)
  
  # 2nd, 4th and 6th visit
  abundance_groupB <- for_plot %>%
    filter(visit_number %in% c(2,4,6)) %>%
    mutate(abundance_groupB = abundance_pc)
  
  # create a single dataframe
  abund_data_group <- full_join(abundance_groupA, abundance_groupB) %>%
    replace_na(list(abundance_groupA = 0, abundance_groupB = 0)) 

  # regression
  plots[[i]] <- ggplot(abund_data_group, mapping = aes(x=abundance_groupA,y=abundance_groupB)) +  geom_point(shape = 21, colour = "black", fill = "white", size = 2, stroke = 1)+ 
  geom_smooth(method="lm", se=TRUE, fullrange=FALSE, level=0.95,linetype="solid") +  
  theme_bw() +
  stat_regline_equation(label.x.npc = "left", aes(label = ..eq.label..),
                        size = 8) +
  stat_regline_equation(label.x.npc = "middle", aes(label = ..rr.label..),
                        size = 8) +
  labs(y="Abundance data from the 2nd, 4th, and 6th visits", 
       x="Abundance data from the 1st, 3rd, and 5th visits",
       title = a) +
  theme(text = element_text(family = "Century Gothic", size = 18, face = "bold"),plot.title = element_text(family = "Century Gothic",
      size = 18, face = "bold"),
      plot.subtitle = element_text(family = "Century Gothic", 
      size = 15, face = "bold",color="#1b2838"),
      axis.title = element_text(family = "Century Gothic",
      size = 18, face = "bold"))
  
  # to extract the R2 value
  df <- data.frame(r_squared = summary(lm(abundance_groupB ~ abundance_groupA,    data = abund_data_group))$r.squared,
                      common_name = a,
                      analysis = "abundance_repeatability")
  # save r2 values
  r2 <- bind_rows(r2, df)
}

# save the r2 values and the plots
cairo_pdf(
  filename = "figs/abundance_vs_abundance_repetability.pdf",
  width = 13, height = 12,
  onefile = TRUE
)
plots
dev.off() 

write.csv(r2, "results/abundance_vs_abundance_repeatabilityR2.csv", row.names = F)
```

## Run analysis at the level of each species (across all sites and treatments combined) for acoustic data 

Here we run analyses separately for point count data and acoustic data. First, we will filter data by visits (choosing the 1st, 3rd and 5th visit) to a site and calculate total acoustic detection rate for each species. We will repeat the above by choosing the 2nd, 4th, and 6th visit to a site and calculate total abundance for each species. 

```{r}
# subset data
# we have 46 species in total
aruData <- aruRate %>%
  filter(common_name %in% spp_subset$common_name)

# creating empty objects to save outputs
plots <- list()
r2_aru <- data.frame()

for (i in 1:length(unique(aruData$common_name))){

 # extract species common name
  a <- unique(aruData$common_name)[i]
  
  # subset aruData for plotting
  for_plot <- aruData[aruData$common_name==a,]
  
  # grouping aruData data
  # 1st, 3rd, and 5th visit
  aru_groupA <- for_plot %>%
    filter(visit_number %in% c(1,3,5)) %>%
    mutate(aru_groupA = aruRate)
  
  # 2nd, 4th and 6th visit
  aru_groupB <- for_plot %>%
    filter(visit_number %in% c(2,4,6)) %>%
    mutate(aru_groupB = aruRate)
  
  # create a single dataframe
  aru_data_group <- full_join(aru_groupA, aru_groupB) %>%
    replace_na(list(aru_groupA = 0, aru_groupB = 0)) 

  # regression
  plots[[i]] <- ggplot(aru_data_group, mapping = aes(x=aru_groupA,y=aru_groupB)) +  geom_point(shape = 21, colour = "black", fill = "white", size = 2, stroke = 1)+ 
  geom_smooth(method="lm", se=TRUE, fullrange=FALSE, level=0.95,linetype="solid") +  
  theme_bw() +
  stat_regline_equation(label.x.npc = "left", aes(label = ..eq.label..),
                        size = 8) +
  stat_regline_equation(label.x.npc = "middle", aes(label = ..rr.label..),
                        size = 8) +
  labs(y="Acoustic detection rate from the 2nd, 4th, and 6th visits", 
       x="Acoustic detection rate from the 1st, 3rd, and 5th visits",
       title = a) +
  theme(text = element_text(family = "Century Gothic", size = 18, face = "bold"),plot.title = element_text(family = "Century Gothic",
      size = 18, face = "bold"),
      plot.subtitle = element_text(family = "Century Gothic", 
      size = 15, face = "bold",color="#1b2838"),
      axis.title = element_text(family = "Century Gothic",
      size = 18, face = "bold"))
  
  # to extract the R2 value
  df <- data.frame(r_squared = summary(lm(aru_groupB ~ aru_groupA,    
                                          data = aru_data_group))$r.squared,
                      common_name = a,
                      analysis = "acousticDetectionRate_repeatability")
  # save r2 values
  r2_aru <- bind_rows(r2_aru, df)
}

# save the r2 values and the plots
cairo_pdf(
  filename = "figs/acousticDetectionRate_vs_acousticDetectionRate_repetability.pdf",
  width = 13, height = 12,
  onefile = TRUE
)
plots
dev.off() 

write.csv(r2, "results/acousticDetectionRate_vs_acousticDetectionRate_repetabilityR2.csv", row.names = F)
```

## Regressions between abundance and acoustic detection rates  

Since this analysis was previously done, we reload the dataframe to get r-squared values between abundance and acoustic detection rates. 
```{r}
# load regressions between aruRates and abundance
regressions <- read.csv("results/aruRates-abundance-regressions.csv")
names(regressions)[6] <- "r_sq_abundance_aruRate"

# load regressions from repeatability analysis (pointCounts)
names(r2) <- c("r_squared_abundance_repeatability", "common_name", "analysis")

# load regressions from repeatability analysis (acousticDetectionRates)
names(r2_aru) <- c("r_squared_aruRate_repeatability","common_name","analysis")

# combine the dataframes
data_regressions <- left_join(regressions[,c(1,6)],r2[,c(1,2)]) %>%
  left_join(., r2_aru[,c(1,2)])

# saving file
write.csv(data_regressions,"results/r_squared_repeatability_regressions.csv",
          row.names = F)
```



## Regressions between abundance from point counts and acoustic detection rates

```{r}

```
